fix: rename module name to `exp`
