--- src/pamiq_vision_exp/__init__.py
@@ -1,3 +1,0 @@
-from importlib import metadata
-
-__version__ = metadata.version(__name__.replace("_", "-"))
