+++ tests/models/components/test_patch_embedding.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from pamiq_vision_exp.models.components.patch_embedding import PatchEmbedding
+from exp.models.components.patch_embedding import PatchEmbedding
 
 
 class TestPatchEmbedding:
